updated the deals remove 50% deals ribbon from two pages

+++ db/migrate/20211124070926_add_is_mobile_in_exit_deliveries.rb
@@ -1,8 +1,9 @@
 class AddIsMobileInExitDeliveries < ActiveRecord::Migration[6.0]
   def change
-    remove_column :exit_deliveries, :is_mobile, :boolean, default: false
-    remove_column :exit_deliveries, :source_rule, :integer, :default => 0
-
+  	if Rails.env.production?
+	    remove_column :exit_deliveries, :is_mobile, :boolean, default: false
+	    remove_column :exit_deliveries, :source_rule, :integer, :default => 0
+  	end
     add_column :exit_deliveries, :is_mobile, :boolean, default: false
     add_column :exit_deliveries, :source_rule, :integer, :default => 0
   end
